Refactor editor state to include core state
This commit refactors the `EditorState` struct to contain a
`CoreEditorState` struct, which contains the core editor state
variables. This change improves code organization and encapsulation.
Also, it moves the handle and clipboard variables into the
`CoreEditorState` struct. Finally, this commit updates the references to
those variables within the rest of the `EditorState` methods.